fast-api-darkside/pyproject.toml
[tool.poetry]
name = "fast-api-darkside"
version = "0.1.0"
description = ""
authors = ["gzhhggg <keiichimonoo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.2"
fastapi-admin = "^1.0.4"
tortoise-orm = "^0.20.0"
aiomysql = "^0.2.0"
aerich = "^0.7.2"
email-validator = "^2.1.0.post1"
uvicorn = "^0.27.0.post1"
faker = "^23.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
